+++ contracts/token/CrunaProtectedNFT.sol
@@ -121,7 +121,7 @@ abstract contract CrunaProtectedNFT is
    * Be careful, your function minting by tokenId MUST check that the tokenId is
    * not higher than nextTokenId. If not, when trying to mint tokens by amount, as soon as
    * nextTokenId reaches the minted tokenId, the function will revert, blocking any future minting.
-   * If you code may risk so, set a function that allow you to correct the nextTokenId to skip
+   * If your code may risk so, set a function that allow you to correct the nextTokenId to skip
    * the token minted by mistake.
    * @param maxTokenId_ The maximum tokenId that can be minted (it can be 0 if no upper limit)
    */
@@ -348,7 +348,7 @@ abstract contract CrunaProtectedNFT is
    * @notice Specify if the caller can call some function.
    * Must be overridden to specify who can manage changes during initialization and later
    * @param isInitializing If true, the function is being called during initialization, if false,
-   * it is supposed to the called later. A time controlled NFT can allow the admin to call some
+   * it is supposed to be called later. A time controlled NFT can allow the admin to call some
    * functions during the initialization, requiring later a standard proposal/execition process.
    */
   function _canManage(bool isInitializing) internal view virtual;