+++ contracts/token/CrunaProtectedNFTTimeControlled.sol
@@ -52,7 +52,7 @@ abstract contract CrunaProtectedNFTTimeControlled is CrunaProtectedNFT, Timelock
    * @notice Specify if the caller can call some function.
    * Must be overridden to specify who can manage changes during initialization and later
    * @param isInitializing If true, the function is being called during initialization, if false,
-   * it is supposed to the called later. A time controlled NFT can allow the admin to call some
+   * it is supposed to be called later. A time controlled NFT can allow the admin to call some
    * functions during the initialization, requiring later a standard proposal/execition process.
    */
   function _canManage(bool isInitializing) internal view virtual override {