+++ contracts/utils/ISignatureValidator.sol
@@ -59,7 +59,7 @@ interface ISignatureValidator {
 
   /**
    * @dev This function validates a signature trying to be as flexible as possible.
-   * As long as called inside the same contract, the cost adding some more parameters is negligible.
+   * As long as called inside the same contract, the cost of adding some more parameters is negligible.
    * Instead, calling it from other contracts can be expensive.
    * @param selector The selector of the function being called.
    * @param owner The owner of the token.